requirements.txt
Pillow==8.3.1
numpy==1.21.1